+++ lib/api.dart
@@ -488,9 +488,7 @@ class InvenTreeAPI {
       List<String> perms = List.from(roles[role] as List<dynamic>);
       return perms.contains(permission);
     } catch (error, stackTrace) {
-      if (error is CastError) {
-        // Ignore CastError
-      } else if (error is TypeError) {
+      if (error is TypeError) {
         // Ignore TypeError
       } else {
         // Unknown error - report it!